README.md
# Titanic-Dataset-Kaggle
This is a simple approach to solve the <a href="https://www.kaggle.com/c/titanic">Titanic: Machine Learning From disaster</a>

Pre-requisites:
1. Python 2/3
2. Scikit-Learn
3. Pandas 
4. Numpy 

I used Patsy's dmatrices function to create formula for training dataset.The problem is, it does not support Scikit-Learn's 
cross_validation() function. So I had to do it manually by splitting the dataset randomly into train and test part.(80% for train 
and 20% for test). 

The submitted result got the score 0.76555.

For improving it, you can follow this <a href="http://ahmedbesbes.com/how-to-score-08134-in-titanic-kaggle-challenge.html"> link</a>.